+++ Sources/Tutti/Onboarding/Onboarding.swift
@@ -10,16 +10,21 @@ import Foundation
 
 /**
  This is a base class for various onboarding experiences. It
- can be subclassed to create more specialized behaviors.
+ can be used as is or subclassed to create different kind of
+ onboarding experiences.
  
- The base onboarding behavior is that `shouldBePresented` is
- `true` until `registerPresentation()` is called. After that,
- it's `false` until `reset()` is called.
- 
- `NOTE` Nothing stops your app from presenting an onboarding
- where `shouldBePresented` is `false`. Make sure that you do
- verify that `shouldBePresented` is `true` and that you call
- `registerPresentation()` when you present it.
+ The basic onboarding behavior is that ``shouldBePresented``
+ is `true` until the first presentation is registered, after
+ which it becomes `false`. The onboarding must then be reset
+ with ``reset()`` before you can present it again.
+
+ You can call ``registerPresentation()`` to register when an
+ onboarding has been presented, if you handle a presentation
+ manually. This will persist this presentation state for the
+ onboarding and make sure that it isn't presented again. You
+ can also call ``tryPresent(presentAction:)`` to perform the
+ provided action if the onboarding should be presented. This
+ causes the presentation to be automatically registered.
  */
 open class Onboarding {
     
@@ -28,8 +33,8 @@ open class Onboarding {
      
      - Parameters:
         - id: The unique onboarding id.
-        - userId: An optional user id.
-        - defaults: The `UserDefaults` instance to use.
+        - userId: An optional user id, by default `nil`.
+        - defaults: The `UserDefaults` instance to use, by default `.standard`.
      */
     public init(
         id: String,
@@ -41,10 +46,17 @@ open class Onboarding {
     }
     
     public typealias UserId = String
-    
+
+
+    /// The unique onboarding id.
     public let id: String
+
+    /// An optional user id.
     public let userId: UserId?
+
+    /// The `UserDefaults` instance to use.
     public let defaults: UserDefaults
+
 
     /**
      Whether or not the onboarding has been presented.
@@ -55,37 +67,44 @@ open class Onboarding {
     
     /**
      Whether or not the onboarding should be presented.
-     
-     This depends on the particular onboarding experience.
+
+     This behavior can be overridden by the onboarding types.
      */
     open var shouldBePresented: Bool {
         presentationCount == 0
     }
     
     /**
-     Call this to register that an onboarding is presented.
+     Register an onboarding presentation.
+
+     This is also called when ``tryPresent(presentAction:)``
+     successfully performs a presentation action.
+
+     Registering a presentation has different effects on the
+     various types. The base ``Onboarding`` and a few of its
+     subclasses types will not be presented again until they
+     are reset, while others behave differently.
      */
     open func registerPresentation() {
         presentationCount += 1
     }
     
     /**
-     Call this to reset any previous presentation logic.
+     Reset the onboarding state.
      */
     open func reset() {
         presentationCount = 0
     }
     
     /**
-     Try presenting the onboarding.
-     
-     This operation will be aborted if the onboarding should
-     not be presented, otherwise the provided `presentAction`
-     block will be called.
+     Try to presenting the onboarding with a certain `action`.
+
+     The presentation attempt is aborted when the onboarding
+     should not be presented.
      */
-    public func tryPresent(presentAction: () -> Void) {
+    open func tryPresent(presentAction action: () -> Void) {
         guard shouldBePresented else { return }
-        presentAction()
+        action()
         registerPresentation()
     }
 }